+++ R/concept_lattice.R
@@ -55,7 +55,7 @@ ConceptLattice <- R6::R6Class(
 
       }
 
-      # Create the SparseConcepts
+      # Create the Concepts
       if (!is.null(extents)) {
 
         private$concepts <- .matrix_to_concepts(
@@ -263,9 +263,9 @@ ConceptLattice <- R6::R6Class(
     #' @description
     #' Get Concepts by Index
     #'
-    #' @param indices (numeric or logical vector) The indices of the concepts to return as a list of SparseConcepts. It can be a vector of logicals where \code{TRUE} elements are to be retained.
+    #' @param indices (numeric or logical vector) The indices of the concepts to return as a list of Concepts. It can be a vector of logicals where \code{TRUE} elements are to be retained.
     #'
-    #' @return A list of SparseConcepts.
+    #' @return A list of Concepts.
     #'
     #' @export
     `[` = function(indices) {
@@ -296,7 +296,7 @@ ConceptLattice <- R6::R6Class(
     #' @param ... See Details.
     #'
     #' @details
-    #' As argument, one can provide both integer indices or \code{SparseConcepts}, separated by commas. The corresponding concepts are used to generate a sublattice.
+    #' As argument, one can provide both integer indices or \code{Concepts}, separated by commas. The corresponding concepts are used to generate a sublattice.
     #'
     #' @return
     #' The generated sublattice as a new \code{ConceptLattice} object.
@@ -387,7 +387,7 @@ ConceptLattice <- R6::R6Class(
     #' @description
     #' Decompose a concept as the supremum of meet-irreducible concepts
     #'
-    #' @param C A list of \code{SparseConcept}s
+    #' @param C A list of \code{Concept}s
     #' @return
     #' A list, each field is the set of meet-irreducible elements whose supremum is the corresponding element in \code{C}.
     #'
@@ -433,7 +433,7 @@ ConceptLattice <- R6::R6Class(
     #' @param ... See Details.
     #'
     #' @details
-    #' As argument, one can provide both integer indices or \code{SparseConcepts}, separated by commas. The corresponding concepts are used to compute their supremum in the lattice.
+    #' As argument, one can provide both integer indices or \code{Concepts}, separated by commas. The corresponding concepts are used to compute their supremum in the lattice.
     #'
     #' @return
     #' The supremum of the list of concepts.
@@ -459,7 +459,7 @@ ConceptLattice <- R6::R6Class(
     #' @param ... See Details.
     #'
     #' @details
-    #' As argument, one can provide both integer indices or \code{SparseConcepts}, separated by commas. The corresponding concepts are used to compute their infimum in the lattice.
+    #' As argument, one can provide both integer indices or \code{Concepts}, separated by commas. The corresponding concepts are used to compute their infimum in the lattice.
     #'
     #' @return
     #' The infimum of the list of concepts.
@@ -482,7 +482,7 @@ ConceptLattice <- R6::R6Class(
     #' @description
     #' Subconcepts of a Concept
     #'
-    #' @param C (numeric or \code{SparseConcept}) The concept to which determine all its subconcepts.
+    #' @param C (numeric or \code{Concept}) The concept to which determine all its subconcepts.
     #'
     #' @return
     #' A list with the subconcepts.
@@ -511,7 +511,7 @@ ConceptLattice <- R6::R6Class(
     #' @description
     #' Superconcepts of a Concept
     #'
-    #' @param C (numeric or \code{SparseConcept}) The concept to which determine all its superconcepts.
+    #' @param C (numeric or \code{Concept}) The concept to which determine all its superconcepts.
     #'
     #' @return
     #' A list with the superconcepts.
@@ -539,7 +539,7 @@ ConceptLattice <- R6::R6Class(
     #' @description
     #' Lower Neighbours of a Concept
     #'
-    #' @param C (\code{SparseConcept}) The concept to which find its lower neighbours
+    #' @param C (\code{Concept}) The concept to which find its lower neighbours
     #'
     #' @return
     #' A list with the lower neighbours of \code{C}.
@@ -568,7 +568,7 @@ ConceptLattice <- R6::R6Class(
     #' @description
     #' Upper Neighbours of a Concept
     #'
-    #' @param C (\code{SparseConcept}) The concept to which find its upper neighbours
+    #' @param C (\code{Concept}) The concept to which find its upper neighbours
     #'
     #' @return
     #' A list with the upper neighbours of \code{C}.
@@ -678,7 +678,7 @@ ConceptLattice <- R6::R6Class(
 
       sets <- sapply(dots,
                      function(l)
-                       inherits(l, "SparseConcept"))
+                       inherits(l, "Concept"))
       sets <- dots[which(sets)]
 
       indices <- sapply(dots, is.numeric)