+++ R/holds_respects.R
@@ -40,12 +40,12 @@
 }
 
 
-#' Check if SparseSet or FormalContext respects an ImplicationSet
+#' Check if Set or FormalContext respects an ImplicationSet
 #'
-#' @param set    (list of \code{SparseSet}s, or a \code{FormalContext}) The sets of attributes to check whether they respect the \code{ImplicationSet}.
+#' @param set    (list of \code{Set}s, or a \code{FormalContext}) The sets of attributes to check whether they respect the \code{ImplicationSet}.
 #' @param imps  (\code{ImplicationSet}) The set of implications to check.
 #'
-#' @return A logical matrix with as many rows as \code{SparseSet}s and as many columns as implications in the \code{ImplicationSet}. A \code{TRUE} in element (i, j) of the result means that the i-th \code{SparseSet} respects the j-th implication of the \code{ImplicationSet}.
+#' @return A logical matrix with as many rows as \code{Set}s and as many columns as implications in the \code{ImplicationSet}. A \code{TRUE} in element (i, j) of the result means that the i-th \code{Set} respects the j-th implication of the \code{ImplicationSet}.
 #'
 #' @export
 #' @importFrom stringr str_pad str_length
@@ -57,7 +57,7 @@
 #' fc %respects% imps
 `%respects%` <- function(set, imps) {
 
-  if (inherits(set, "SparseSet")) {
+  if (inherits(set, "Set")) {
 
     set <- list(set)
 
@@ -68,7 +68,7 @@
     S <- lapply(seq_along(set$objects),
                   function(i) {
 
-                    SparseSet$new(attributes = set$attributes,
+                    Set$new(attributes = set$attributes,
                                   M = set$I %>% extract_columns(i))
 
                   })
@@ -78,7 +78,7 @@
 
   if (inherits(set, "list") &&
       (length(set) > 0) &&
-      inherits(set[[1]], "SparseSet")) {
+      inherits(set[[1]], "Set")) {
 
     res <- lapply(set, function(s) .respect(s, imps)) %>%
       purrr::reduce(cbind)