+++ tests/testthat/test-sparse_set.R
@@ -1,4 +1,4 @@
-context("SparseSet")
+context("Set")
 
 test_that("fcaR operates on sparse sets", {
 
@@ -32,15 +32,15 @@ test_that("fcaR operates on sparse sets II", {
 
 })
 
-test_that("fcaR uses class SparseSet", {
+test_that("fcaR uses class Set", {
 
   attributes <- paste0("P", 1:6)
 
-  expect_error(A <- SparseSet$new(attributes = attributes), NA)
+  expect_error(A <- Set$new(attributes = attributes), NA)
   expect_error(A$assign(attributes = "P1", values = 0.3), NA)
 
   expect_error(A["P1"], NA)
-  expect_is(A["P1"], "SparseSet")
+  expect_is(A["P1"], "Set")
   expect_equal(A["P1"]$cardinal(), 0.3)
 
   expect_is(A$get_vector(), "SpM")
@@ -56,7 +56,7 @@ test_that("fcaR uses class SparseSet", {
   expect_true(A %<=% A)
 
   expect_error(v <- as_vector(A), NA)
-  expect_error(A2 <- as_SparseSet(v), NA)
+  expect_error(A2 <- as_Set(v), NA)
 
   expect_true(A %==% A2)